Address comments and add onClick
